+++ src/Resources/InboundDocumentResource.php
@@ -6,16 +6,22 @@ namespace Deinte\ScradaSdk\Resources;
 
 use Closure;
 use Deinte\ScradaSdk\Dto\InboundDocument;
+use Deinte\ScradaSdk\Exceptions\AuthenticationException;
 use Deinte\ScradaSdk\Exceptions\NotFoundException;
+use Deinte\ScradaSdk\Exceptions\ScradaException;
+use Deinte\ScradaSdk\Exceptions\ValidationException;
 use Deinte\ScradaSdk\Requests\InboundDocuments\ConfirmDocumentRequest;
 use Deinte\ScradaSdk\Requests\InboundDocuments\GetDocumentPdfRequest;
 use Deinte\ScradaSdk\Requests\InboundDocuments\GetDocumentRequest;
 use Deinte\ScradaSdk\Requests\InboundDocuments\GetUnconfirmedDocumentsRequest;
 use Deinte\ScradaSdk\Resources\Concerns\HandlesResponseErrors;
+use Deinte\ScradaSdk\ScradaConnector;
 use Saloon\Http\BaseResource;
 
 /**
  * Inbound document operations.
+ *
+ * @property ScradaConnector $connector
  */
 final class InboundDocumentResource extends BaseResource
 {
@@ -25,10 +31,16 @@ final class InboundDocumentResource extends BaseResource
      * Fetch all unconfirmed inbound documents.
      *
      * @return array<int, InboundDocument>
+     *
+     * @throws AuthenticationException
+     * @throws ValidationException
+     * @throws ScradaException
      */
     public function getUnconfirmed(): array
     {
-        $response = $this->connector->send(new GetUnconfirmedDocumentsRequest);
+        $response = $this->connector->send(new GetUnconfirmedDocumentsRequest(
+            $this->connector->getCompanyId()
+        ));
 
         $this->throwIfError($response);
 
@@ -38,23 +50,30 @@ final class InboundDocumentResource extends BaseResource
             return [];
         }
 
-        $items = array_filter(
-            $data,
-            static fn (mixed $item): bool => is_array($item)
-        );
+        $result = [];
+        foreach ($data as $item) {
+            if (is_array($item)) {
+                $result[] = InboundDocument::fromArray($item);
+            }
+        }
 
-        return array_map(
-            static fn (array $item): InboundDocument => InboundDocument::fromArray($item),
-            array_values($items)
-        );
+        return $result;
     }
 
     /**
      * Fetch a single inbound document by ID.
+     *
+     * @throws AuthenticationException
+     * @throws NotFoundException
+     * @throws ValidationException
+     * @throws ScradaException
      */
     public function get(string $documentId): InboundDocument
     {
-        $response = $this->connector->send(new GetDocumentRequest($documentId));
+        $response = $this->connector->send(new GetDocumentRequest(
+            $this->connector->getCompanyId(),
+            $documentId
+        ));
 
         $this->throwIfError($response, $this->notFoundFactory($documentId));
 
@@ -69,10 +88,18 @@ final class InboundDocumentResource extends BaseResource
 
     /**
      * Download the PDF for an inbound document.
+     *
+     * @throws AuthenticationException
+     * @throws NotFoundException
+     * @throws ValidationException
+     * @throws ScradaException
      */
     public function getPdf(string $documentId): string
     {
-        $response = $this->connector->send(new GetDocumentPdfRequest($documentId));
+        $response = $this->connector->send(new GetDocumentPdfRequest(
+            $this->connector->getCompanyId(),
+            $documentId
+        ));
 
         $this->throwIfError($response, $this->notFoundFactory($documentId));
 
@@ -81,10 +108,18 @@ final class InboundDocumentResource extends BaseResource
 
     /**
      * Confirm that a document has been handled.
+     *
+     * @throws AuthenticationException
+     * @throws NotFoundException
+     * @throws ValidationException
+     * @throws ScradaException
      */
     public function confirm(string $documentId): bool
     {
-        $response = $this->connector->send(new ConfirmDocumentRequest($documentId));
+        $response = $this->connector->send(new ConfirmDocumentRequest(
+            $this->connector->getCompanyId(),
+            $documentId
+        ));
 
         $this->throwIfError($response, $this->notFoundFactory($documentId));